lab1 ex6 修复了page开头表示slot使用情况的byte数组长度计算错误的问题

+++ src/java/simpledb/storage/HeapPage.java
@@ -80,7 +80,7 @@ public class HeapPage implements Page {
      * @return the number of bytes in the header of a page in a HeapFile with each tuple occupying tupleSize bytes
      */
     private int getHeaderSize() {        
-        return (int)Math.ceil(this.numSlots/8);
+        return (int)Math.ceil((double)this.numSlots/8);
     }
     
     /** Return a view of this page before it was modified
